+++ src/App.tsx
@@ -13,12 +13,12 @@ import AdminOnlyPrivate from 'auth/AdminOnlyPrivate';
 
 import HomePage from 'pages/HomePage';
 import KakaoAuthPage from 'pages/KakaoAuthPage';
-import SignupPage from 'pages/SignupPage/SignupPage';
-import InvitedPage from 'pages/alba-InvitedPage/InvitedPage';
-import AddGroupPage from 'pages/admin-AddGroupPage/AddGroupPage';
-import SelectWeekPage from 'pages/admin-SelectWeekPage/SelectWeekPage';
-import ApplicationOpenPage from 'pages/admin-ApplicationOpenPage/ApplicationOpenPage';
-import ApplicationClosePage from 'pages/admin-ApplicationClosePage/ApplicationClosePage';
+import SignupPage from 'pages/SignupPage';
+import InvitedPage from 'pages/alba/InvitedPage';
+import AddGroupPage from 'pages/admin/AddGroupPage';
+import SelectWeekPage from 'pages/admin/SelectWeekPage';
+import ApplicationOpenPage from 'pages/admin/ApplicationOpenPage';
+import ApplicationClosePage from 'pages/admin/ApplicationClosePage';
 
 function App(): JSX.Element {
   return (
