--- src/components/PageStyledComponents/admin-ApplicationOpenPage/setPeople/index.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const InputPeople = styled.input`
-  margin-left: auto;
-  width: 4rem;
-  background-color: transparent;
-  font-size: ${({ theme }) => theme.fonts.fontSize.xl};
-`;
