+++ src/components/PageStyledComponents/admin/ApplicationClose/index.tsx
