+++ src/hooks/admin-ApplicationOpenPage/useTimeTemplate.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { getTimeTemplate } from 'apis/adminApplication';
 import { useAtom, useAtomValue, useSetAtom } from 'jotai';
-import { openStepAtom, timeTemplateAtom, weeklyPeopleAtom } from 'pages/admin-ApplicationOpenPage/ApplicationOpenPage';
+import { openStepAtom, timeTemplateAtom, weeklyPeopleAtom } from 'pages/admin/ApplicationOpenPage';
 import React, { useEffect } from 'react';
 import weekdayArray from 'utils/weekdayArray';
