+++ src/pages/alba/InvitedPage/InvitationContent.tsx
